feat: workspace policy

+++ app/Http/Controllers/WorkspaceController.php
@@ -14,6 +14,7 @@ use App\Http\Resources\CardResource;
 use Illuminate\Http\RedirectResponse;
 use App\Http\Requests\WorkspaceRequest;
 use App\Http\Resources\WorkspaceResource;
+use Illuminate\Support\Facades\Gate;
 
 class WorkspaceController extends Controller
 {
@@ -72,6 +73,8 @@ class WorkspaceController extends Controller
 
     public function edit(Workspace $workspace): Response
     {
+        Gate::authorize('update_workspace', $workspace);
+
         return inertia(component: 'Workspaces/Setting', props: [
             'workspace' => fn() => new WorkspaceResource($workspace->load('members')),
             'page_settings' => [
@@ -86,6 +89,8 @@ class WorkspaceController extends Controller
 
     public function update(Workspace $workspace, WorkspaceRequest $request): RedirectResponse
     {
+        Gate::authorize('update_workspace', $workspace);
+
         $workspace->update([
             'name' => $name = $request->name,
             'slug' => str()->slug($name . str()->uuid(10)),
@@ -114,6 +119,8 @@ class WorkspaceController extends Controller
 
     public function member_store(Workspace $workspace, Request $request): RedirectResponse
     {
+        Gate::authorize('member_workspace', $workspace);
+
         $request->validate([
             'email' => [
                 'required',
@@ -144,6 +151,8 @@ class WorkspaceController extends Controller
 
     public function member_destroy(Workspace $workspace, Member $member): RedirectResponse
     {
+        Gate::authorize('delete_workspace', $workspace);
+
         $member->delete();
 
         flashMessage('Member successfully deleted');
